+++ Scripts/Services/IAttackType.cs
@@ -1,4 +1,4 @@
 public interface IAttackType
 {
-    public int CalculateDamage(DefenceComponent defenceComponent, float damage);
+    public float CalculateDamage(DefenceComponent defenceComponent, float damage);
 }
